% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_.R
\name{bossier_example}
\alias{bossier_example}
\title{Get path to example data}
\usage{
bossier_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
bossier has example data in \code{inst/extdata}
directory. This function eases access (from jennybc - readxl)
}
\examples{
bossier_example()
bossier_example("bossier_example.sav")
}
