% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_.R
\name{pad_q_num}
\alias{pad_q_num}
\title{Pad SPSS .sav question codes}
\usage{
pad_q_num(text, pad)
}
\arguments{
\item{text}{the question name}

\item{pad}{number of zeros to pad at start of question number}
}
\value{
Returns a zero padded question id
}
\description{
Pad SPSS .sav question codes
}
