% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_.R
\name{read_bos}
\alias{read_bos}
\title{Import spss .sav data from Online Surveys (BOS)}
\usage{
read_bos(filename, sep = "---")
}
\arguments{
\item{filename}{path to SPSS .sav file from BOS}

\item{sep}{separator for question code and question text}
}
\value{
Returns a tibble.
}
\description{
Import spss .sav data from Online Surveys (BOS)
}
